train_classifier_AL

+++ train_classifier_AL_1.py
@@ -177,8 +177,8 @@ def get_dataloader(bgf,sigf, batch_size=32, shuffle=False, num_workers=4):
 
     idx = np.random.permutation(len(label))
     
-    jet1 = torch.tensor(jet1_data[idx], dtype=torch.float32)
-    jet2 = torch.tensor(jet2_data[idx], dtype=torch.float32)
+    jet1 = torch.tensor(jet1_data[idx], dtype=torch.int64)
+    jet2 = torch.tensor(jet2_data[idx], dtype=torch.int64)
     padding_mask1 = torch.tensor(padding_mask1[idx], dtype=torch.bool)
     padding_mask2 = torch.tensor(padding_mask2[idx], dtype=torch.bool)
     label = torch.tensor(label[idx], dtype=torch.float32)  # For BCEWithLogitsLoss, labels should be float32
